Move Utterances to surfaces
